backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymssql==2.2.11
cx_Oracle==8.3.0
openai==1.12.0
pinecone-client==3.0.2
langchain==0.1.7
langchain-openai==0.0.5
pandas==2.2.0
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
numpy==1.26.3
scipy==1.12.0
notion-client==2.2.1
markdownify==0.11.6
requests==2.31.0
python-multipart==0.0.6
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0